[#736] CLI: allow "--ignore-standalone-config" flag to override CSV values (#928)
RepoSense ignores the --ignore-standalone-config command line flag when
parsing a repo-config csv file.

We can allow the --ignore-standalone-config command line flag to
override all entries in the "Ignore Standalone Config" column of
repo-config.csv. This provides users with an easy method of setting
this flag for all repos, without having to edit each line of the csv.

To achieve this, let's move the isStandaloneConfigIgnored attribute
from the LocationsCliArguments class, to its superclass: CliArguments.
This will allow us to access the --ignore-standalone-config flag's
value when parsing RepoConfigurations from both LocationsCliArguments
and ConfigCliArguments.