Fix additional upward movement on expanding code (#966)
Clicking the expand button may cause the code to move up a bit 
before expanding. This occurs more frequently when the number of 
code lines to load is large, or the client is slow. 

This is due to asynchronous data loading by Vue, causing observable 
CSS updates to the code block before the code has been loaded into 
the DOM.

Let's execute the CSS change for the code block only after the 
segment's code lines have been loaded.