[#900] group-config.csv: make location optional (#914)
In the group-config.csv, location is a compulsory field that only has one
value.  Specifying groups for all repos requires the user to manually key
in groups for each repo.

However, this can be improved such that if the location field is not
filled, we can bind that GroupConfig to all the RepoConfig. This can
help the user save the time listing all the possible combinations of
repos and groups when they want to specify groups for all repos.

Let's bind the GroupConfig to all the RepoConfig if the location
field is not filled.