[#876] FileType: default file type to "other" if file format is not of standard convention (#877)
The label of a FileType only allows alphanumeric characters.

When format is not specified, files without a standard file format
and with special characters (e.g.: "$gradle_backup$", "Backup File")
will fail the FileType validation test and RepoSense terminates.

Let's:
* Allow FileType label to accept any non-empty string
* Enforce file format validation to be alphanumeric
* Label the FileType to "other" where file format cannot be deduced
  for files with the unorthodox naming convention
* Ensure that each file exists in the host computer before
  attempting to analyze them.

A file format will be classified as "other" if the string of the
format portion of the file name contains non-alphanumeric
characters, e.g.: "main.j@va". (+2 squashed commits)
Squashed commits:
[2b2a9f3] [#785] ChartView: fix authors with unaccounted lines (#789)

The git blame command used in RepoSense contains the option -w, which
ignores whitespace when comparing the parent’s version and the child’s
to find where the lines came from.

When an author did not change the content but only the position of
specific lines of code, the previous author will be recognised as the
author that changed these lines. If the previous author did not
contribute during the analysis period, his/her ramp chart will be
empty while the contribution bar will be present.

Let's add an extra check in aggregateBlameInfo for lines of code with
the author committed date that is not within the duration of
specified since date and until date. Those lines of code's author will
be set to unknown, which will resolve the issue above.
[fda0acf] [#831] Extra white space in code panel in Microsoft Edge (#974)

We set the height attribute of the active segments to 100% to expand
the code.

In Chrome, because the height of the parent node of this segment is
not set, Chrome will ignore this attribute and set it auto. However,
in Edge, it is actually setting the height of the segment to be the
height of the window, leaving extra white space.

Let's explicitly set the height of active segments to auto to fix
the problem.