[#831] Extra white space in code panel in Microsoft Edge (#974)
We set the height attribute of the active segments to 100% to expand
the code.

In Chrome, because the height of the parent node of this segment is
not set, Chrome will ignore this attribute and set it auto. However,
in Edge, it is actually setting the height of the segment to be the
height of the window, leaving extra white space.

Let's explicitly set the height of active segments to auto to fix
the problem.