[#785] ChartView: fix authors with unaccounted lines (#789)
The git blame command used in RepoSense contains the option -w, which
ignores whitespace when comparing the parent’s version and the child’s
to find where the lines came from.

When an author did not change the content but only the position of
specific lines of code, the previous author will be recognised as the 
author that changed these lines. If the previous author did not
contribute during the analysis period, his/her ramp chart will be
empty while the contribution bar will be present.

Let's add an extra check in aggregateBlameInfo for lines of code with
the author committed date that is not within the duration of 
specified since date and until date. Those lines of code's author will
be set to unknown, which will resolve the issue above.