[#852] Zoom tab displays commits not within filter time rage (#971)
When we try to open the zoom tab for the same user with the same
date range for the second time, commits displayed the second time
contain commit that does not fall within the filter time range.

This is mainly due to we edit the user object of info property
directly. However, info is the property of the child component so
we are not supposed to change it directly. The situation mentioned
above occurs because this property is overwritten again by the
parent, which related to the way we re-render the sidebar.

Let's use another computed value to represent the user with filtered
commits so we do not need to modify the property directly. Minor
changes are done together to improve code quality.