[#181] Inconsistent thickness of different contribution bars (#976)
The height of the contribution bar appears to be different for different
contributors.

It is because we use rem to represent the height, which may result
in a non-integer px value. And the browsers may display the same
floating-point px differently.

Let's make it look consistent for every contributor by explicitly
setting the height of the pixel to be in px.