[#1009] AppVeyor Failing: remove outdated tests (#1015)
AppVeyor CI is failing in all PRs as it has updated its worker 
image, which uses Git 2.25 instead of Git 2.23.

From Git v2.24, Git for windows would print error for an illegal
file name and prevent it from adding into the index. As in our test 
cases we run the command `git checkout someInvalidName`, an 
exception is expected to be thrown, and since we do not catch it, 
the tests fail. 

Let's remove the 2 outdated test as the git command now handles
detection of invalid filename and will throw exception at the very
beginning.