[#885] Allow reset search filter (#895)
The user has to manually delete the keywords and press another enter 
to go back to the full view.

However, this can be improved by providing a reset button to clear 
the search bar, which can help the user save time for deleting. 

Let's provide a button to reset the search bar.