[#1001] CSS: Increase z-index for center divider div (#1011)
When we scroll horizontally in the Code Panel, the code snippets and 
code toggle buttons overlap with the center divider.

Let's fix this bug by increasing the z-index of the #tab-resize div, and 
the div that closes the right tab. We can use a value of 100, since the 
z-index of #tab-resize-guide div is already set to 100. For better 
scalability, let's also use a SASS map to store these z-index values.