[#988] Ramp chart: inconsistency in the length of contribution line (#991)
Toggling the `breakdown by filetype` checkbox changes the length of the
contribution bar even though the LoC remains constant.

This occurs due to the inaccurate data in json file. The
authorFileTypeContributionMap in commit.json have multiple entries of the 
same key. In the frontend, the last entry would be taken instead of all 
of the entries provided. This results in the wrong calculation of the total 
contribution of each file type.

Let's override FileType#hashCode to correctly update the LoC number 
instead of creating a new entry.

In the backend, we use AuthorshipSummary#addAuthorContributionCount
to increment line number and we consider FileType (the key) and
LoC (the value) as an entry in the HashMap. The HashMap is then output
to the commit.json file as authorFileTypeContributionMap. However, the 
default HashCode will digest all data stored in FileType, other data 
like path and pathsGlob is also taken into consideration when comparing 
the key to update the entry. So, instead of updating, a new entry will 
be created. Thus in AuthorshipSummary#authorFileTypeContributionMap, 
each author can have multiple entry with same FileType#label. And the 
resulted authorFileTypeContributionMap in json file will have multiple 
entry of the same key.