Remove duplicate commits in daily contributions
Two or more authors under with the same extracted commit may result in
the commit being added multiple times to an author’s daily contribution.

This occurs when these authors use some identical aliases, which causes
the same author to be retrieved from the commit result for adding to
the daily contribution.
This can be reproduced by configuring said authors, or using no author
config under the above conditions.

Let's add a duplicate commit check to the commit analyzing process to
prevent this.