[#1032] Author: fix error in copy constructor (#1036)
The Author class has a constructor that accepts an existing Author
object and creates a copy. 

However, the constructor assigns the list of author aliases of the 
copied instance to the `ignoreGlobList` attribute of the new 
instance.

Let's fix this error so methods that use this constructor will not
face erroneous behavior.