[#920] Ramp chart: show percentile for each item (#922)
We have index number for each group in the ramp chart but not its
percentile. It is hard to tell the ranking position of the group.

This can be improved to show the percentile of each item so that it
is easy to see whether an item is in the 20%th percentile and so on.

Let's add a percentile number at the right side of each group.