Improve file collapsing in authorship view
The authorship tab relies on several dom querying methods to toggle
whether a file is collapsed from view.
This hurts the performance of file collapsing, and detracts from the
use of vue as RepoSense’s front-end framework, decreasing code
readability.

Let's refactor this to use a new variable in the file object instead,
allowing vue reactivity to take over this behaviour.
This also fixes a bug with files not staying collapsed after changing
the file filter [#1081].

With this change, we are able to conveniently display both the expand
and collapse all buttons at the same time. Let's do so as well.

Occasionally, a small number of files have a large number of line
contributions. This slows the initial loading time at best, and causes
the application to crash at worst.
Let's use the new variable to collapse files with more than
2000 lines of code by default, alleviating this.