[#999] Remove empty zoom__day divs (#1000)
When granularity is set to week, the 'pushCommitsWeek' method gets 
all possible weeks (week 1, 2, 3, etc...) from start date until the 
end date and pushes it to 'user.commits', no matter the week has any 
commits at all. Additionally, commits with zero insertions but with 
deletions are also not taken into consideration in the commits 
panel.

Weeks that have no commits is not useful and redundant. Commits with
only deletions are maybe useful as it may represent 
refactoring/removing redundant code.

Let's add a check to add a week only if it has commits. And remove 
the check for commits with zero insertions, so that commits with 
only deletions are also included.