[#906, #994, #1006] UserGuide: update doc and remove errors (#907)
With the recently added group-config.csv file, the new changes 
possible in the local configuration are not reflected in the
documentation.

Users can configure their csv files to analyse multiple repos in 
their own custom way, but they may be confused about how to do so as 
the user guide is outdated. 

Let's update the user guide to reflect the changes possible through
group-config.csv in the local configuration to analyse multiple 
repos. Let's also fix the errors in the Tool Bar section and fix the 
formatting of repo-config.csv example.