[#1023] Authorship Panel: expand all button always appear when first open the page (#1040)
The `expand all` button will always appear on the authorship panel 
when we first open the page or when refresh the page, until the
loading is complete. The button to display is supposed to be
`collapse all` as we are to display all the file details initially.

Let's display the `collapse all` button at the very beginning, and 
update all relevant checkboxes, radio buttons using proper Vue 
model.