[#1071] Zoom view: provide a way to highlight the commit (#1075)
Clicking a tag jumps to a list of commits full of tags. Users may
not be able to find the commit with that tag easily.

We can highlight the resulting commit with that particular tag by 
providing a light blue border around the message div.

Let's add a focus property to the commit message body.