[#1043] Fix bug of incorrect commit dates when merge group (#1045)
When merge group is checked, some commits are seen with committed
dates that are incompatible to the real committed dates.

Commits with the same date are appended together in 'mergedCommits'.
Since 'mergedCommits' is being sorted before all the merged commits 
for a particular group is gathered, the order of the 'mergedCommits' 
is changed, and some commits with the same date are appended 
wrongly.

Let's move the sorting of 'mergedCommits' after the all merged 
commits are gathered, this will ensure commits with the same date 
are appended correctly.

The order of the 'mergedCommits' matters, as we use an additional js 
object 'dateToIndexMap' maps dates to the corresponding index in 
'mergedCommits'. Once the order has changed, the 'dateToIndexMap' no 
longer contains the accurate mappings and hence, commits are 
appended wrongly.