[#1044] Update date hashes on reset date range (#1068)
Resetting the date range would occasionally not change the date 
hashes in the URL, causing a refresh of the page to load the 
outdated filter date range. 

This bug occurs when until or since date is not specified on
generating the report.

Let's make sure that the date hashes gets updated all the time.