[#1005, #1010] Code Panel: match font with GitHub and break lines when overflow (#1008)
When the file path or lines of code overflows, it takes up a large 
amount of empty space in order to accomodate to it's length. The 
font size of the code and file path is also rather big and takes up 
much space.

Let's set the file path and lines of code to break into next lines 
when it overflows. Also, let's set the font-family and font-size of 
the code to match with GitHub.