Expand pug-lint glob and remove fix option in lint
The pug-lint command does not expand globs provided as arguments,
relying instead on bash wildcard expansion. This causes the pug-lint
command to fail on windows.
The main linting script was also written with the --fix option, which
can increase the build time for CI.
This may also cause auto-fixable errors to go undetected, although the
lint script has an extra ‘--' that causes --fix to be recognised as a
file.

Let’s manually expand the globs for the pug-lint command, moving it to
a separate script for brevity.
Let’s also remove the --fix option used by the main linting script,
and move it to a separate lintfix script.
At the same time, let’s introduce --fix for stylelint as well.