[#1145] SideBar: revamp display logic (#967)
When sidebar needs to be refreshed, such as re-rendering the 
authorship tab and zoom tab, we set `isTabActive` to false first to 
trigger the redrawing process. To ensure that the child component 
gets the updated changes, we bind a unique key to the child 
component to trigger re-rendering.

The previous implementation is done in such a way because the child 
component fails to get the changes in `info` prop due to the way we 
update the `info`[1]. This causes the SideBar component to be 
completely re-rendered, instead of only updating the component based 
on the changes in `info`. If we are able to let the SideBar 
component catch the change of `info`, we can directly watch the 
property and update the data upon the `info` change.

Let's declare the `info` prop explicitly and watch the changes in 
`info` prop so Vue is able to catch the property change[2] to 
utilize components re-usage. Let's also pass a shallow copied `user` 
to the zoom tab, as we are intending to make zoom tab and summary 
tab independent of one another.

[1]https://vuejs.org/v2/guide/reactivity.html#Change-Detection-Caveats
[2]https://vuejs.org/v2/guide/reactivity.html#Declaring-Reactive-Properties


Co-authored-by: jamessspanggg <jamespangmunwai@gmail.com>