Reposense class: throw exception earlier
Currently, `Reposense` continues with report generation even though
there are no valid repo locations in the `--repo` arguments.

This results in a late crash during the execution of the
`cloneAndAnalyzeRepos` method quite some time after Reposense runs.

Let's
Add a state-test at the end of the `getRepoConfigurations` method, to
check if any repo location(s) are valid. If there are none, then throw
an `IllegalStateException`.

`getRepoConfigurations` is the method that constructs a list of
`RepoLocation` objects from the inputs. It is the earliest way to check
if there are valid locations, making it easier to identify the cause of
the command failure. `IllegalStateException` is thrown as it is not
possible for the application to recover and generate reports, because
the precondition that is at least 1 valid argument is not met.