[#1124] Authorship: file name starting with period not displayed (#1125)
File with name starting with period will not have its detail 
displayed on authorship panel.

This is due to the minimatch library used which by default will not
match name `a/**/b` with `a/.d/b`, unless dot is set. So the file is
filtered out from selected files.

Let's set the dot property to true to allow the file details to be
displayed properly.