[#823, #981] AnnotatorAnalyzer: Fix code disowning mechanism (#1067)
Suppose that `random-user` is not the git author name of any author
in a git repository(i.e. it is just a placeholder name). RepoSense 
does not correctly mark code as disowned when a user marks a code 
block with the `@@author random-user` tag. Only the line containing 
the `@@author random-user` tag is treated as disowned, while the 
remaining code block is still attributed to the author deduced from 
the git blame analysis.

This bug occurs because the code in AnnotatorAnalyzer uses a 
constant UNKNOWN_AUTHOR to represent both the case where there is no 
currently active annotation override AND the case where the user has 
used a "@@author random-user" tag.

Let's use Optionals to store the author names obtained from the
`@@author` tags, so we can identify the case where the user has used
a `@@author random-user` tag.