Save authorship tab controls as hash parameters
The filter and sorting options for the authorship tab are not persisted
across page reloads.

This results having to reselect the options whenever the page is
reloaded, leading to a less pleasant user experience.

Let's save the options for these controls as hash parameters, and use
them across page reloads when available.