[#1083] RepoConfiguration: remove `annotationOverwrite` flag  (#1117)
There is an `annotationOverwrite` flag in the `RepoConfiguration`. 
The flag is redundant because it defaults to `true` and is never 
utilized in the main code.

Since there is currently no planned use case, for better code 
quality, we should remove the unnecessary code.

Let's remove this flag from the `RepoConfiguration`.