[#1046] v_summary: refactor duplicate code (#1056)
There is an instance of code duplication in 'v_summary.js'. 
Specifically, the definition of a function that is passed to the 
comparator has been duplicated in the case of grouping by repos and 
by authors.

Let's extract this duplicated portion of code into a helper function 
so that it may be referenced by the comparator.