[#873] Enable view commits option when merge group (#1100)
The view commits option is not available when merge group option is 
checked.

Let's enable the view option when merge group, as the zoom in 
function is already working with merged ramps. Let's also fix the 
regression from #1106 that removes the github icon when merge group 
and `filterGroupSelection` is 'groupByRepos'.