[#202, #1081] Authorship: improve file collapsing mechanism (#1065)
The authorship tab relies on several dom querying methods to toggle
whether a file is collapsed from view. This hurts the performance of 
file collapsing and detracts from the use of Vue as RepoSense’s 
front-end framework, decreasing code readability.

Occasionally, a small number of files have a large number of line 
contributions. This slows the initial loading time at best and 
causes the application to crash at worst.

Let's
- add a new variable 'active' in the `file` object, allowing Vue 
  reactivity to take over this toggle behavior, which also fixes 
  the bug with files not staying collapsed after changing the file 
  filter.
- display both the expand and collapse all buttons according to that
  new property.
- collapse files with more than 2000 lines of code by default.