[#1147] Generate random filetype colour when predefined colours are used up (#1148)
The filetype colours are chosen through a rotation of the predefined 
colour values. So if the number of filetypes exceeds the number of 
predefined colours, there will be repetitions of the same colour.

The repeated filetypes may be hard to distinguish among one
another, in terms of their colours.

Let's generate a random colour for the filetypes after the standard
selection of predefined colours have been exhausted.