[#1181] Allow filtering of ramps by file type (#1192)
All ramps are shown even when some file types are filtered out, the
sizes of ramps are also not adjusted when their filetypes are 
filtered out.

This may cause confusion in users as they would think that the ramps 
will be filtered out accordingly.

Let's add the filtering of ramps by the file types that are selected, 
the size of the ramp will also be adjusted based on that.

Additional Info:
1. The current git log results does not include filetypes that are
modified, so I decided to get from the backend by changing the
git log format. The filetype details of each commit are also added
to the commit.json

2. 'git log --numstat --shortstat' is used instead of 'git log --stat'
to retrieve the log with files changed. This is because --stat only 
shows the sum of lines changed in file whereas --numstat displays 
addition and deletion separately. Addition and deletion of binary
files will be '-' and '-'.

3. When removing a filetype from a ramp, the corresponding addition
and deletion will be removed from the ramp. A ramp will only be 
removed when all of it's filetypes are removed.