[#1213] Regressions from #1059 (#1218)
With the recent merged PR #1059, the data related to sorting were not
passed from v_summary to v_summary_charts, even though
v_summary_charts uses those data within openTabZoom method.

This results in zoom tab to not be restorable at some situations, as it 
is unable to utilise the sorting options. However, the sorting options
and the call for sortFiltered is actually not needed
within the restoreZoomFiltered method. This is because the
matchZoomUser within restoreZoomFiltered will already filter out
other groups/repos. Thus, there is no need to pass those props into
v_summary_charts.

Let's remove the unnecessary parameters and methods.