[#1178] Make github profile and repo hyperlinks more consistent (#1186)
When grouped by repo, the same icon to link to the repo is duplicated
multiple times for no reason. Also, the icon to link to the author's
profile is hard to find.

We can make the links more intuitive and consistent by putting links to
the author's profile beside author names, and links to repo beside repo
names.

Let's reorder the icon group such that the hyperlinks are closer to the
respective names, as well as always showing the icons beside their
respective names. As this change heavily relies on the icons being
responsive, let's resolve issue #1191 as well to make use of
vue-fontawesome to make the icons responsive.