[#1170] Update total contribution to reflect based on checked file types (#1171)
When some filetypes are unchecked, the contribution bar length adjusts 
to the correct length, while the total line contribution of each group 
does not. Additionally, when toggling sorting controls to decreasing 
order of contribution, groups with higher total contribution are seen 
to be in lower position than groups with lower contribution. This may 
cause confusion among users.

Let's update the total contribution to reflect the total contribution 
among those file types that are checked.