[#1242] Move removal of zoom hashes to be within v_zoom component (#1243)
The removal of zoom hashes are done in main.js, while the removal
of authorship hashes is done within the v_authorship component
itself.

We can move the responsibility to remove zoom hashes to be within 
the v_zoom component itself, in the 'beforeDestroy' hook. This 
will result in a cleaner implementation and separation of concern.

Let's move the removal of zoom hashes to be within v_zoom
component itself.