Fix bug when restoring zoom tab (#1261)
When group by authors and groups are merged, the commits are 
not shown correctly in the zoom tab after reloading the page.

This occurs especially for an author with contribution in several 
repositories, as the repositories of the author are treated as 
separate groups (should be pushed into same group) within the
'restoreZoomFiltered' method. This leads to the merging of group 
not being able to merge all the repositories of the author, as 
they are treated as separate groups. 

The searched field is also not handled by zoom tab as well, 
leading to the reloaded zoom tab containing commits that are
not by users that match the searched field.

Let's modify the 'restoreZoomFiltered' method to treat the
repositories of an author to be in the same group, instead of 
being separate groups. Let's also add the handling of the
search field when reloading the zoom tab.