[#1221] Allow merging of individual groups (#1223)
Currently the `merge group` checkbox only supports merging/expanding all 
groups. 

It is inconvenient for users who want to have an overview across all 
repos first, then take a detailed look within a particular group, because
there is a need to relocate the group on the page after expanding all 
groups. This is especially the case when there are many authors and repos
on the dashboard.

Let's add the feature to allow merging/expanding individual groups 
instead.