Revamp docs (#1281)
Revamp docs to leverage MarkBind features

We recently converted docs to MarkBind format. But the content was
mostly translated from the previous format which was in simple
Markdown format.

Let's revamp the docs to take advantage of MarkBind features. As this
is going to touch almost all docs, let's also take this opportunity
to improve the content where possible.

High-level overview of the changes:
* Split all-in-one-pages UG and DG into smaller pages
* Use different siteNav for UG and DG
* Add top-level pages: SHOWCASE, CONTACT
* Beautify landing page, differentiate it from the README
* Use a default layout for all pages
* Cross-embed relevant content in relevant places
* Use dark green as the brand color
* Refer to se-edu/guides for reference materials (e.g., style guides)