Use variables for page titles and headings to reduce duplication
The page title and the main heading of a page is exactly the same.
Let's use a MarkBind variable to hold that value so that it can be
specified only once and used in both as title and heading.

At the same time, let's fix some inconsistencies between page titles
and headings that were partially caused by the duplication this
commit addresses.