Use nunjucks to generate siteNavs and site map in the index page
<include>ing the (ug|dg)SiteNave.md in the landing page site-map avoids
duplication but prevents optimal use of siteNav as well as site-map.
For example, we cannot use :expanded: in the siteNav to make an item
collapsible because :expanded: appears as raw text in the landing page.

Let's use nunjucks to define siteNav items as an array and a macro to
render it optimally for the siteNav and landing page.

Pros:
* siteNav items are now collapsible
* landing page shows links as a hierarchy rather than a flat list

Cons:
* The code is more complicated than before: But, the complexity is
  within the expertise of RepoSense developers as nunjucks code is
  similar to JavaScript.