Use variables for page titles and headings to reduce duplication (#1291)
The page title and the main heading of a page is exactly the same.
Let's use a MarkBind variable to hold that value so that it can be
specified only once and used in both as title and heading.

At the same time, let's fix some inconsistencies between page titles
and headings that were partially caused by the duplication this
commit addresses.

Warning: Two pages (runSh.md, usingAuthorTags.md) do not use
{{ title }} inside the <frontmatter>. This is due to a flaw in the
current MarkBind processor that affects variable values that
have `.` and `@` in them. This flaw has been fixed in the next
upcoming MarkBind release. Those two pages need to be 
updated after that version of MarkBind has been released.