UG/DG: Fix the current item in the siteNav not being marked (#1298)
UG/DG: Fix the current item in the siteNav not being marked

The site navigation menu of the UG and DG does not highlight the
currently loaded page in blue, as it is supposed to.
The reason is, the {{baseUrl}} variable used in the macro does not
get imported with the macro. The links generated still work because
the baseUrl is empty in this case. But they don't get highlighted due
to some strictness in how MarkBind finds which item to highlight.

Let's fix it by adding `with context` to the nunjucks import statement.