Update Gradle Version (#1300)
The `downloadNode` Gradle extension is outdated, as a result, CI is 
failing. 

Let's update the Gradle version to keep up to date with its 
development.

From Gradle 5.0 onwards, the tasks running with code coverage are 
configured to delete the execution data before they start executing
another task. As a result, our code coverage dropped significantly 
since we use the same file to store the execution data of the 
different tests. 

Let's also use different files to store the execution data of the 
different tests to fix this problem.