[#1240] Remove insertion and deletion of each commit in commits.json (#1269)
Since the line contribution info of each file type is already within 
individual commit, we can remove the insertion and deletion info of 
each commit in the `commits.json` file. Same for
`authorFinalContributionMap`.

The insertion and deletion info can be calculated in the front
end instead, by summing up all the insertion and deletion of
all filetypes within a commit.

Let's remove the insertion and deletion in commit.json file and
calculate them in the front end.