[#1263] Tests: Simplify resource loading (#1264)
Tests: Simplify code used for loading resources

In many of our tests, we make repeated use of the ClassLoader 
to obtain the path to each resource used in the tests. 
This has resulted in a lot of repetitive code.

Let's create a helper method to reduce this repetition,
and simplify the resource loading code.