[#1129] RepoSense: exit gracefully when all repo locations are invalid (#1157)
RepoSense invokes `ReportGenerator::generateReposReport`
even if all repository locations in the arguments are invalid.

This is due to the try-catch block in
`RepoSense::getRepoConfigurations`, which prevents the
`InvalidLocationException` to be thrown and interrupt
the execution of the program. Therefore, RepoSense cannot abort 
early when all repository locations are invalid.

Let's add one more condition check before return, such that
a `ParseException` will be thrown if the RepoConfiguration list is 
empty. By catching that exception, RepoSense can exit gracefully.