[#1232] Fix footer link (#1305)
The link in the footer of the RepoSense report is outdated as we have moved
everything to the website.

Let's update the links in the footer to point to the website based on whether 
a specific version or master is being used for deployment. 

If the master branch is used for deployment, we point to the latest UG. If the 
latest release is used, we point to the UG in the production site.