[#1257] AuthorConfigCsvParser: Initialize author fully before adding to config (#1270)
The `AuthorConfigCsvParser` adds each `Author` object to the
`AuthorConfiguration` before the author is fully initialized
with its details (emails, aliases, display name, ignore glob list).

This requires the `AuthorConfiguration` to be updated again after 
the `Author` object is fully initialized. 

Let's initialize the `Author` object fully before adding it into the
`AuthorConfiguration` object. With this change, we can also remove
the code that re-updates the author config after initializing the 
`Author` object.