[#155] Display date ranged commit in side bar (#1101)
When the granularity is set to daily or weekly, clicking on the ramp 
slice redirect the user to a new tab where they will see the 
information of the individual commits from GitHub itself.

This requires the user to switch between different tabs to view 
relevant info. This can be distracting and troublesome for the user,
especially when trying to traverse through multiple weeks of 
commits. We can improve this by opening the zoom tab for the 
corresponding date ranged commits instead of open another Github 
page.

Let's open the zoom tab when the user clicking the ramp with the
granularity set to daily or week, instead of opening new page.