[#1322] Fix loading page rendering issue (#1334)
The contents within `v-resizer` show at the beginning of page load.

This is because we write our code directly to `index.pug` instead of
a separate file, e.g. app.pug. The generated HTML file from 
`index.pug` contains some body HTML code, which is recognized by the 
browser and displayed first before the JS code is loaded.

Let's move the v-if conditional out of v-resizer to a Vue template,
since the Vue template tag is not recognizable by the browser.