[#1312] Conditional run for markbind to gh pages deployment (#1337)
The markbind to gh pages deployment task of Travis is executed at 
every pull request, which runs npm install but do not deploy to the
gh pages since it is not the master branch. This is unnecessary 
and takes up build time.

Let's ensure that the deployment task to be only run when it is from 
the master branch.