[#1205] Show a tooltip when hovering over ambiguous icons and text (#1331)
The report has existing tooltips that are rendered by the browser. 
However, the behaviour is not consistent across different browsers 
and in certain cases, not appear at all.

Let's introduce a CSS-based functionality for showing tooltips to
the user and provide a quick explanation on the usage of some of the
report features.