Generate the report based on the user's timezone
The report is not currently configured to respect the --timezone
configuration setting, which may cause issues where a commit is
incorrectly attributed to the wrong date in terms of the user's
local time zone.

Let's introduce a functionality that makes the report generation
more aware about the --timezone configuration setting, such that
any date and time related functionality is made timezone-aware.