Show non-blank line count in the brakets instead of blank lines
Currently, the report in the code panel (authorship) part shows
the number of blank lines written by an author in a file or in all
files of a file type inside the brackets beside the total number of
lines written. The viewer needs to do mental calculation to get the
number of non-blank lines written by the author.

Let's change the front end code to show non-blank line count instead.