Allow specifying a range of commits to ignore
Currently, each commit that needs to be ignored has to be
specified accordingly, one entry at a time. This may be
troublesome for use cases where a range of commits has to be
ignored.

Let's introduce a functionality that allows users to specify a
range of commits using the Git ".." notation, so as to ignore
commits that are within the given two commits inclusive in the
report generation.